Ignore upgrades to node.
Because of issues in Node 8.14, we do not want to raise pull requests for upgrading the version of node in our projects.

We will deal with upgrades as part of a transformation across all repositories.

This is only until we are happy that our systems are stable on the lastest versions of node again.

Resolves https://github.com/Financial-Times/next/issues/280.