Don't raise patch PRs for internal dependencies
Add an exception in for UI specific repos which we will always want to
merge quickly between systems.

There is a feeling that renovate is opening too many PRs, there isn't
generally a need to keep everything up to date to the nearest patch
level so don't automatically open PRs for patch level releases of
internal packages. This doesn't stop us from manually creating PRs for
those if we do want them.